un/signed compares

+++ trunk/src/arp-ioctl.c
@@ -291,7 +291,7 @@ arp_loop(arp_t *r, arp_handler callback, void *arg)
 		if (msg.len >= sizeof(*tea) && tea->PRIM_type == T_ERROR_ACK)
 			return (-1);
 		
-		if (rc != MOREDATA || msg.len < sizeof(*toa) ||
+		if (rc != MOREDATA || msg.len < (int)sizeof(*toa) ||
 		    toa->PRIM_type != T_OPTMGMT_ACK ||
 		    toa->MGMT_flags != T_SUCCESS)
 			return (-1);
